Use libtool dependencies for linking tests
If libservicelog is not already installed on the system library path, the
test suite currently fails because the library cannot be found.  This also
means that the library being tested by the test suite is *not* the one in
the source tree, but the one installed on the system.

This is fixed by linking against the libtool .la instead of directly against
the .so, so that we get a wrapper script that sets the LD_LIBRARY_PATH.

Signed-off-by: Steve Langasek <steve.langasek@canonical.com>
Signed-off-by: Ankit Kumar <ankit@linux.vnet.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>