Output alignment issues in sl_printf
Commit c7dd78a fixed the output alignment issue in "v29_print.c". The same commit is getting
backported to the servicelog.c and also taking care the issues mentioned below.

If there were two consecutive newline characters, the snprintf would be called with length
set to 1. The snprintf returns the destination string length instead of the actual length written
over the buffer, which breaks the logic. To get rid of such issues, the fixes introduced to write
the newline explicitly and adjust the buf_offset.

Signed-off-by: Anup Kumar <anupkumk@linux.vnet.ibm.com>
Reviewed-By: Shivaprasad G Bhat <sbhat@linux.ibm.com>
[Removed changelog from description - Vasant]
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>